requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
Flask==2.0.0
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
joblib==1.0.1
MarkupSafe==2.0.0
nltk==3.6.2
regex==2021.4.4
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
tqdm==4.60.0
urllib3==1.26.4
Werkzeug==2.0.0
wincertstore==0.2
newsapi-python
scikit-learn